feat: 데이터 베이스 동적 연결 설정 로직

+++ src/lib/crud.py
@@ -2,8 +2,12 @@ from lib.databases import Databases
 import psycopg2
 import psycopg2.extras  # psycopg2.extras 모듈을 임포트
 
+
 class Crud(Databases):
     def insertDB(self, table, column, data):
+        if not self.cursor:
+            print("Database not connected")
+            return
         sql = "INSERT INTO {} ({}) VALUES %s;".format(table, column)
         try:
             self.cursor.execute(sql, (data,))
@@ -11,24 +15,17 @@ class Crud(Databases):
         except Exception as e:
             print("Insert DB error:", e)
 
-    def insertDBSeq(self, table, column, data, seq):
-        sql = "INSERT INTO {} ({}) VALUES %s;".format(table, column)
-        try:
-            self.cursor.execute(sql, (data,))
-            self.db.commit()
-        except Exception as e:
-            print("Insert DB error:", e)
-
-    def readDB(self, table, column,return_column_names=False):
-        
+    def readDB(self, table, column, return_column_names=False):
+        if not self.cursor:
+            print("Database not connected")
+            return []
         sql = f"SELECT {column} FROM {table}"
-        
         try:
             self.cursor.execute(sql)
             result = self.cursor.fetchall()
             if return_column_names:
                 colnames = [desc[0] for desc in self.cursor.description]
-                return result , colnames
+                return result, colnames
             else:
                 return result
         except Exception as e:
@@ -36,27 +33,36 @@ class Crud(Databases):
             raise e
 
     def whereDB(self, table, column, where, return_column_names=False):
+        if not self.cursor:
+            print("Database not connected")
+            return []
         sql = "SELECT {} FROM {} WHERE {}".format(column, table, where)
-        try: 
+        try:
             self.cursor.execute(sql)
             result = self.cursor.fetchall()
             if return_column_names:
                 colnames = [desc[0] for desc in self.cursor.description]
-                return result , colnames
+                return result, colnames
             else:
                 return result
         except Exception as e:
             print("Read DB error:", e)
 
     def updateDB(self, schema, table, column, value, condition):
+        if not self.cursor:
+            print("Database not connected")
+            return
         sql = "UPDATE {}.{} SET {}=%s WHERE {}=%s".format(schema, table, column, column)
         try:
             self.cursor.execute(sql, (value, condition))
             self.db.commit()
         except Exception as e:
             print("Update DB error:", e)
-    
+
     def deleteDB(self, table, condition):
+        if not self.cursor:
+            print("Database not connected")
+            return
         sql = "DELETE FROM {} WHERE {};".format(table, condition)
         try:
             self.cursor.execute(sql)
@@ -65,30 +71,23 @@ class Crud(Databases):
             print("Delete DB error:", e)
 
     def execute_param_query(self, query, params):
-        """
-        파라미터화된 쿼리를 실행하고 결과를 반환하는 메서드
-        :param query: 실행할 쿼리 문자열
-        :param params: 파라미터로 대체할 값들이 담긴 딕셔너리
-        :return: 쿼리 결과
-        """
+        if not self.cursor:
+            print("Database not connected")
+            return []
         try:
-            # ':' 파라미터를 실제 값으로 대체
             formatted_params = {key.lstrip(':'): value for key, value in params.items()}
             for key, value in formatted_params.items():
                 query = query.replace(f":{key}", f"{{{key}}}")
-
-            # `params` 딕셔너리의 키를 `{}`로 둘러싸인 플레이스홀더에 맵핑하여 대체
             query = query.format(**formatted_params)
-
-            print("Executing Query:", query)  # 쿼리 확인용 로그
-
+            print("Executing Query:", query)
             self.cursor.execute(query)
             return self.cursor.fetchall()
         except Exception as e:
             print("Query execution with parameters error:", e)
             return []
 
-if __name__ == "__main__":
-    db = Crud()
-    print(db.readDB(table='public.my_table', column='*'))
-    db.updateDB(schema='public', table='my_table', column='name', value='new_value', condition="id = 1")
+
+# if __name__ == "__main__":
+#     db = Crud()
+#     print(db.readDB(table='public.my_table', column='*'))
+#     db.updateDB(schema='public', table='my_table', column='name', value='new_value', condition="id = 1")
